requirements.txt
# Core dependencies
weave>=0.50.0
python-dotenv>=1.0.0
pyyaml>=6.0
numpy>=1.24.0

# LLM providers
openai>=1.35.0
anthropic>=0.30.0
google-generativeai>=0.5.0

# UI and visualization
rich>=13.7.0
click>=8.1.0

# Async support
aiohttp>=3.9.0
httpx>=0.25.0
asyncio-throttle>=1.0.2

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Optional performance
orjson>=3.9.0  # Faster JSON parsing